// src/app/shared/models/manageConsultation.ts
export interface manageConsultation {
    consultation_id: string;
    consultation_date: string;
    consultation_time: string;
    consultation_consultor: string;
    consultation_consultee: string;
    consultation_type: string;
    consultation_platform: string;
    consultation_comments: string;
    consultation_status: string;
}